Add display images module
